% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{adstock_weibull_revised}
\alias{adstock_weibull_revised}
\title{\link{Suggestion-5b} Adstock Weibull New}
\usage{
adstock_weibull_revised(
  x,
  shape,
  scale,
  windlen = length(x),
  type = "cdf",
  index_end = NULL
)
}
\arguments{
\item{x}{TODO::Type. TODO::Description}

\item{shape}{TODO::Type. TODO::Description}

\item{scale}{TODO::Type. TODO::Description}

\item{windlen}{= length(x) TODO::Type. TODO::Description}

\item{type}{= "cdf" TODO::Type. TODO::Description}

\item{index_end}{= Null}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
