% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{adstock_weibull_revised}
\alias{adstock_weibull_revised}
\title{\emph{\strong{Suggestion}} adstock_weibull_revised}
\usage{
adstock_weibull_revised(
  x,
  shape,
  scale,
  windlen = length(x),
  type = "cdf",
  index_end = NULL
)
}
\arguments{
\item{x}{Numeric.}

\item{shape}{Numeric.}

\item{scale}{Numeric. Check "Details" section for more details.}

\item{windlen}{Integer. Length of modelling window. By default, same length as \code{x}.}

\item{type}{Character. Accepts "CDF" or "PDF". CDF, or cumulative density
function of the Weibull function allows changing decay rate over time in both
C and S shape, while the peak value will always stay at the first period,
meaning no lagged effect. PDF, or the probability density function, enables
peak value occurring after the first period when shape >=1, allowing lagged
effect.}

\item{index_end}{Numeric. Makes scales have a fixed value generated from
the training period}
}
\value{

}
\description{
Newly defined the adstock_weibull_revised, which receive
an additional argument index_end that makes scales have
a fixed value generated from the training period.
}
