% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocator.R
\name{get_allocator_benchmarks}
\alias{get_allocator_benchmarks}
\title{\emph{\strong{Suggestion}} Allocator Results}
\usage{
get_allocator_benchmarks(
  AllocatorCollect_opt,
  AllocatorCollect_hist,
  AllocatorCollect_recent
)
}
\arguments{
\item{AllocatorCollect_opt}{TODO::Type. TODO::Description}

\item{AllocatorCollect_hist}{TODO::Type. TODO::Description}

\item{AllocatorCollect_recent}{TODO::Type. TODO::Description}
}
\value{
List(). Contains list of hyperparameters.
}
\description{
TODO:: Write description
Generate analytical result of allocato
}
