% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{Allocator_results_new}
\alias{Allocator_results_new}
\title{\emph{\strong{Suggestion}} Allocator_results_new}
\usage{
Allocator_results_new(
  InputCollect,
  AllocatorCollect_opt,
  AllocatorCollect_hist,
  AllocatorCollect_recent
)
}
\arguments{
\item{AllocatorCollect_opt}{TODO::Type. TODO::Description}

\item{AllocatorCollect_hist}{TODO::Type. TODO::Description}

\item{AllocatorCollect_recent}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
