% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{budget_boundary}
\alias{budget_boundary}
\title{\link{Suggestion-2} Boundary}
\usage{
budget_boundary(InputCollect, cost_input_low, cost_input_high, total_spends)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{cost_input_low}{TODO::Type. TODO::Description}

\item{cost_input_high}{TODO::Type. TODO::Description}

\item{total_spends}{TODO::Type. TODO::Description}
}
\value{
List(). Contains list of boundaries.
}
\description{
TODO:: Write description
(originally) robyn channel boundary
need to modify the function into the genral form of the Robyn
library
}
