% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{compare_nonzero}
\alias{compare_nonzero}
\title{\emph{\strong{Suggestion}} compare_nonzero}
\usage{
compare_nonzero(InputCollect, AllocatorCollect, pre_date)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{pre_date}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
\emph{\strong{Suggestion}}: Able to analyze without non-zero spend days]
focus on all days (w/ non-zero spend days)
check accuracy with dependent
}
