% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{compare_zero_spend}
\alias{compare_zero_spend}
\title{\emph{\strong{Suggestion}} compare_zero_spend}
\usage{
compare_zero_spend(InputCollect, AllocatorCollect, pre_date)
}
\description{
focus on all days (w/ non-zero spend days)
check accuracy with dependent
}
