% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{decompose_dependent_vars}
\alias{decompose_dependent_vars}
\title{\emph{\strong{Suggestion}} decompose_dependent_vars}
\usage{
decompose_dependent_vars(InputCollect, pre_period = NULL)
}
\description{
Decomposing the dependent variable into paid media channels and the rest
}
\details{
Returns the average weekly (or daily) total dependent variables
on paid media channels and the rest subtracted from
the average weekly value of the dependent variable.
}
