% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{generate_budget_boundaries}
\alias{generate_budget_boundaries}
\title{\emph{\strong{Suggestion}} generate_generate_budget_boundaries}
\usage{
generate_budget_boundaries(
  InputCollect,
  cost_input_low,
  cost_input_high,
  total_spends
)
}
\value{
List(). Contains list of boundaries.
}
\description{
To help marketers set their lower and upper budget
boundaries in a more intuitive way.
}
