% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocator.R
\name{get_allocator_benchmarks}
\alias{get_allocator_benchmarks}
\title{\emph{\strong{Suggestion}} get_allocator_benchmarks}
\usage{
get_allocator_benchmarks(
  AllocatorCollect_opt,
  AllocatorCollect_hist,
  AllocatorCollect_recent
)
}
\arguments{
\item{AllocatorCollect_opt}{}

\item{AllocatorCollect_hist}{}

\item{AllocatorCollect_recent}{}
}
\value{
List(). Contains list of hyperparameters.
}
\description{
Provides some critical benchmarks to help analysts
to evaluate the efficiency of the suggested optimal allocation.
}
