% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{get_allocator_benchmarks_zero_spend}
\alias{get_allocator_benchmarks_zero_spend}
\title{\emph{\strong{Suggestion}} get_allocator_benchmarks_zero_spend}
\usage{
get_allocator_benchmarks_zero_spend(
  InputCollect,
  AllocatorCollect_opt,
  AllocatorCollect_hist,
  AllocatorCollect_recent
)
}
\description{
\emph{\strong{Suggestion}} get_allocator_benchmarks_zero_spend
}
