% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{get_individual_result_new}
\alias{get_individual_result_new}
\title{\emph{\strong{Suggestion}} Allocator_results_new}
\usage{
get_individual_result_new(
  InputCollect,
  OutputCollect,
  media_metric,
  select_model,
  type = "mean",
  pre_period = NULL
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{media_metric}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}

\item{type}{= "mean"}

\item{pre_period}{= Null}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
