% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{get_response_sum_on_train_zero_spend}
\alias{get_response_sum_on_train_zero_spend}
\title{\emph{\strong{Suggestion}} get_response_sum_on_train_zero_spend}
\usage{
get_response_sum_on_train_zero_spend(InputCollect, OutputCollect, select_model)
}
\arguments{
\item{InputCollect}{}

\item{OutputCollect}{}
}
\value{

}
\description{
TODO:: Write description
}
