% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{get_response_sum_on_trainining}
\alias{get_response_sum_on_trainining}
\title{\emph{\strong{Suggestion}} get_response_sum_on_trainining}
\usage{
get_response_sum_on_trainining(InputCollect, OutputCollect, select_model)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
Returns the sum of the response made by each paid media channel
during the training period.
}
\details{
The ratio of the response from the predict_response_sum_on_test
function to that from the get_response_sum_on_trainining function
indicates the predicted response change from the training period
to the test period.
}
