% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{get_response_sum_on_trainining_new}
\alias{get_response_sum_on_trainining_new}
\title{\emph{\strong{Suggestion}} get_response_sum_on_trainining_new}
\usage{
get_response_sum_on_trainining_new(InputCollect, OutputCollect, select_model)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
