% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{predict_individual_result}
\alias{predict_individual_result}
\title{\emph{\strong{Suggestion}} predict_individual_result}
\usage{
predict_individual_result(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  media_metric,
  select_model,
  type = "mean"
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{type}{= "mean" TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
Predicts the response and dependent variables made by each media
during the test period
}
