% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{predict_individual_result_new}
\alias{predict_individual_result_new}
\title{\emph{\strong{Suggestion}} predict_individual_result_new}
\usage{
predict_individual_result_new(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  media_metric,
  select_model,
  type = "mean"
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{type}{= "mean"}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
form of the post_data is equal to the form of the InputCollect$dt_input
just it has different date range
}
