% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{predict_individual_result_zero_spend}
\alias{predict_individual_result_zero_spend}
\title{\emph{\strong{Suggestion}} predict_individual_result_zero_spend}
\usage{
predict_individual_result_zero_spend(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  media_metric,
  select_model,
  type = "mean"
)
}
\arguments{
\item{type}{= "mean"}
}
\description{
TODO:: Write description
form of the post_data is equal to the form of the InputCollect$dt_input
just it has different date range
}
