% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{predict_response_sum_on_test}
\alias{predict_response_sum_on_test}
\title{\emph{\strong{Suggestion}} predict_response_sum_on_test}
\usage{
predict_response_sum_on_test(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\description{
Predicts the sum of the response made by each media
during the test period
}
