% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{predict_response_sum_on_test_zero_spend}
\alias{predict_response_sum_on_test_zero_spend}
\title{\emph{\strong{Suggestion}} predict_response_sum_on_test_zero_spend}
\usage{
predict_response_sum_on_test_zero_spend(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
