% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{predict_response_sum_on_test_zero_spend}
\alias{predict_response_sum_on_test_zero_spend}
\title{\emph{\strong{Suggestion}} predict_response_sum_on_test_zero_spend}
\usage{
predict_response_sum_on_test_zero_spend(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\description{
TODO:: Write description
}
