% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{put_hyper_params}
\alias{put_hyper_params}
\title{\emph{\strong{Suggestion}} put_hyper_params}
\usage{
put_hyper_params(InputCollect = NULL, vars_lst, hyper_lst)
}
\description{
Concatenates multiple lists, each of which groups channels
sharing the same hyperparameter ranges.
}
