% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{result_media}
\alias{result_media}
\title{\link{Suggestion-6} Get Performance of Each Media}
\usage{
result_media(
  InputCollect,
  OutputCollect,
  media_metric,
  select_model,
  type = "mean",
  pre_period = NULL
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{media_metric}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}

\item{type}{= "mean" TODO::Type. TODO::Description}

\item{pre_period}{= Null}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
