% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{predict_individual_result}
\alias{predict_individual_result}
\title{\link{Suggestion-8}}
\usage{
predict_individual_result(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  media_metric,
  select_model,
  type = "mean"
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{type}{= "mean" TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
\link{Sub function for upper suggestion: make result based on post period}
form of the post_data is equal to the form of the InputCollect$dt_input
ust it has different date range
}
