% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{result_total}
\alias{result_total}
\title{\link{Suggestion-9a} Get Total Performance from whole media}
\usage{
result_total(InputCollect, OutputCollect, select_model)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
