% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{predict_response_sum_on_test}
\alias{predict_response_sum_on_test}
\title{\link{Suggestion-9b} Get Total Performance from whole media}
\usage{
predict_response_sum_on_test(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
