% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{robyn_onepagers_revised}
\alias{robyn_onepagers_revised}
\title{robyn_onepagers_revised}
\usage{
robyn_onepagers_revised(
  InputCollect,
  OutputCollect,
  select_model = NULL,
  quiet = FALSE,
  export = TRUE
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{select_model}{= Null TODO::Type. TODO::Description}

\item{quiet}{= False TODO::Type. TODO::Description}

\item{export}{= True TODO::Type. TODO::Description}
}
\value{
List(). Contains list of hyperparameters.
}
\description{
TODO:: Write description
\link{Suggestion included in the function (63:105)}
Original version is in plot.R
}
