% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{robyn_onepagers_revised}
\alias{robyn_onepagers_revised}
\title{\emph{\strong{Suggestion}} included] robyn_onepagers_revised}
\usage{
robyn_onepagers_revised(
  InputCollect,
  OutputCollect,
  select_model = NULL,
  quiet = FALSE,
  export = TRUE
)
}
\arguments{
\item{select_model}{= Null}

\item{quiet}{= False}

\item{export}{= True}
}
\value{
List().
}
\description{
Suggestion included in (82:121). Original version is in plot.R
Suggestion: Message of Convergence on plot.
Reason: Practical suggestion for checking convergence
on the resulted plots, (not only logs)
}
