% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{saturation_hill_new}
\alias{saturation_hill_new}
\title{\link{Suggestion-5a} Saturation Hill New}
\usage{
saturation_hill_new(x, alpha, gamma, index_end = NULL, x_marginal = NULL)
}
\arguments{
\item{x}{TODO::Type. TODO::Description}

\item{alpha}{TODO::Type. TODO::Description}

\item{gamma}{TODO::Type. TODO::Description}

\item{index_end}{= Null TODO::Type. TODO::Description}

\item{x_marginal}{= Null TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
