% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_predicts}
\alias{validate_predicts}
\title{\emph{\strong{Suggestion}} Validate the results by history and prediction}
\usage{
validate_predicts(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{}

\item{OutputCollect}{}

\item{post_data}{}

\item{post_period}{}

\item{select_model}{}
}
\value{

}
\description{
compare history and prediction
need to modify the output form
}
