% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{validate_predicts_zero_spend}
\alias{validate_predicts_zero_spend}
\title{\emph{\strong{Suggestion}} validate_predicts_zero_spend}
\usage{
validate_predicts_zero_spend(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\description{
TODO:: Write description
compare history and prediction
need to modify the output form
}
