% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_predicts}
\alias{validate_predicts}
\title{\emph{\strong{Suggestion}} Validate the results by history and prediction}
\usage{
validate_predicts(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
compare history and prediction
need to modify the output form
}
