% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{validation_test_new}
\alias{validation_test_new}
\title{\emph{\strong{Suggestion}} validation_test_new}
\usage{
validation_test_new(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
compare history and prediction
need to modify the output form
}
