% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{Allocator_results}
\alias{Allocator_results}
\title{\link{Suggestion-3} Allocator Results}
\usage{
Allocator_results(
  AllocatorCollect_opt,
  AllocatorCollect_hist,
  AllocatorCollect_recent
)
}
\arguments{
\item{AllocatorCollect_opt}{TODO::Type. TODO::Description}

\item{AllocatorCollect_hist}{TODO::Type. TODO::Description}

\item{AllocatorCollect_recent}{TODO::Type. TODO::Description}
}
\value{
List(). Contains list of hyperparameters.
}
\description{
TODO:: Write description
Generate analytical result of allocato
}
