% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_zero.R
\name{get_allocator_benchmarks_zero_spend}
\alias{get_allocator_benchmarks_zero_spend}
\title{\emph{\strong{Suggestion}} get_allocator_benchmarks_zero_spend}
\usage{
get_allocator_benchmarks_zero_spend(
  InputCollect,
  AllocatorCollect_opt,
  AllocatorCollect_hist,
  AllocatorCollect_recent
)
}
\arguments{
\item{AllocatorCollect_opt}{TODO::Type. TODO::Description}

\item{AllocatorCollect_hist}{TODO::Type. TODO::Description}

\item{AllocatorCollect_recent}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
