% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{decompose_dependent_vars}
\alias{decompose_dependent_vars}
\title{\link{Suggestion-4} Decompose KPI of media, non-media, etc}
\usage{
decompose_dependent_vars(InputCollect, pre_period = NULL)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{pre_period}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
check accuracy with dependent
}
