% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{generate_budget_boundaries}
\alias{generate_budget_boundaries}
\title{\emph{\strong{Suggestion}} generate_generate_budget_boundaries}
\usage{
generate_budget_boundaries(
  InputCollect,
  cost_input_low,
  cost_input_high,
  total_spends
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{cost_input_low}{TODO::Type. TODO::Description}

\item{cost_input_high}{TODO::Type. TODO::Description}

\item{total_spends}{TODO::Type. TODO::Description}
}
\value{
List(). Contains list of boundaries.
}
\description{
To help marketers set their lower and upper budget
boundaries in a more intuitive way.
}
