% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{get_individual_result}
\alias{get_individual_result}
\title{\emph{\strong{Suggestion}} get_individual_result}
\usage{
get_individual_result(
  InputCollect,
  OutputCollect,
  media_metric,
  select_model,
  type = "mean",
  pre_period = NULL
)
}
\arguments{
\item{InputCollect}{}

\item{OutputCollect}{}

\item{media_metric}{}

\item{select_model}{}

\item{type}{= "mean" }

\item{pre_period}{= Null}
}
\value{

}
\description{
Returns the response and dependent variables made by
each paid media channel during the training period.
}
