% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{get_individual_result_zero_spend}
\alias{get_individual_result_zero_spend}
\title{\emph{\strong{Suggestion}} get_individual_result_zero_spend}
\usage{
get_individual_result_zero_spend(
  InputCollect,
  OutputCollect,
  media_metric,
  select_model,
  type = "mean",
  pre_period = NULL
)
}
\arguments{
\item{InputCollect}{}

\item{OutputCollect}{}

\item{media_metric}{}

\item{select_model}{}

\item{type}{= "mean"}

\item{pre_period}{= Null}
}
\value{

}
\description{
TODO:: Write description
}
