% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{predict_individual_result}
\alias{predict_individual_result}
\title{\emph{\strong{Suggestion}} predict_individual_result}
\usage{
predict_individual_result(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  media_metric,
  select_model,
  type = "mean"
)
}
\arguments{
\item{InputCollect}{}

\item{OutputCollect}{}

\item{post_data}{}

\item{post_period}{}

\item{type}{= "mean" }
}
\value{

}
\description{
Predicts the response and dependent variables made by each media
during the test period
}
