% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_channel.R
\name{predict_response_sum_on_test}
\alias{predict_response_sum_on_test}
\title{\emph{\strong{Suggestion}} predict_response_sum_on_test}
\usage{
predict_response_sum_on_test(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
Predicts the sum of the response made by each media
during the test period
}
