% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{put_hyppar}
\alias{put_hyppar}
\title{\link{Suggestion-1} Set hyperparameters from Input}
\usage{
put_hyppar(InputCollect = NULL, vars_lst, hyper_lst)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{vars_lst}{TODO::Type. TODO::Description}

\item{hyper_lst}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
