% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{result_total_post}
\alias{result_total_post}
\title{\link{Suggestion-9b} Get Total Performance from whole media}
\usage{
result_total_post(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
}
