% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_predicts}
\alias{validate_predicts}
\title{\emph{\strong{Suggestion}} Validate the results by history and prediction}
\usage{
validate_predicts(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\description{
Presents a table comparing the predicted values from each approach
with the actual values and calculating the errors
}
