% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_spend.R
\name{validate_predicts_zero_spend}
\alias{validate_predicts_zero_spend}
\title{\emph{\strong{Suggestion}} validate_predicts_zero_spend}
\usage{
validate_predicts_zero_spend(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
compare history and prediction
need to modify the output form
}
