% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validation_test}
\alias{validation_test}
\title{\link{Suggestion-10} Validate the results by history and prediction}
\usage{
validation_test(
  InputCollect,
  OutputCollect,
  post_data,
  post_period,
  select_model
)
}
\arguments{
\item{InputCollect}{TODO::Type. TODO::Description}

\item{OutputCollect}{TODO::Type. TODO::Description}

\item{post_data}{TODO::Type. TODO::Description}

\item{post_period}{TODO::Type. TODO::Description}

\item{select_model}{TODO::Type. TODO::Description}
}
\value{
TODO::Type. TODO::Description
}
\description{
TODO:: Write description
compare history and prediction
need to modify the output form
}
